refactor: refactor shared/ui for future purposes
