+++ src/shared/ui/components/error-message/index.ts
