+++ src/shared/ui/components/index.ts
@@ -1,6 +1,5 @@
-export { ErrorBoundary } from './error-boundary';
-export { Loader } from './loader';
-export { Input } from './input';
 export { Button } from './button';
 export { ErrorMessage } from './error-message';
+export { Input } from './input';
+export { Loader } from './loader';
 export { Pagination } from './pagination';
