+++ src/shared/ui/components/loader/index.ts
