+++ src/shared/ui/components/loader/ui/Loader.module.scss
