+++ src/shared/ui/components/loader/ui/Loader.spec.tsx
