+++ src/shared/ui/components/pagination/ui/Pagination.module.scss
