+++ src/shared/ui/providers/error-boundary/index.ts
