+++ src/shared/ui/providers/error-boundary/ui/ErrorBoundary.spec.tsx
