+++ src/shared/ui/providers/index.ts
@@ -1,0 +1,1 @@
+export { ErrorBoundary } from './error-boundary';
